FEM_solver quadratic 1D elements Python
This code is written in Python 2.7.3 using spyder IDE on Ubuntu 12.04 LTS. It is a simple case with a rod of specified thermal properties subjected to 1 dimensional steady state heat conduction with internal heat generation. Heat transfer rate, internal heat generation per unit volume and rod end temperature are the inputs required. The program gives nodal temperatures and a plot of temperature against rod length (obtained using interpolation functions) as output. This is rather a rudimentary code and there is a large potential for minimizing the memory consumption and optimization. This case is solved using the elementary approach.