FEM_solver 4-node quadrilateral 2D elements Python
This code is written in Python 2.7.3 using spyder IDE on Ubuntu 12.04 LTS. It involves 2 dimensional steady state thermal analysis of a rectangular fin subjected to a certain inlet temperature. I have tried to make this code as generic as possible. I have included 2 dimensional heat conduction, internal heat generation along with area and edge heat convection. Fin thermal properties, inlet temperature and ambient temperature are required as input parameters. This code gives nodal temperatures as output. Elimination approach is used for solving this case. The code chooses the number of elements on its own. The fin face subjected to the inlet temperature is assumed to be isothermal.